Rename paypal handlers to controllers
