+++ app/Controllers/PaypalNotificationController.php
@@ -2,7 +2,7 @@
 
 declare( strict_types = 1 );
 
-namespace WMDE\Fundraising\Frontend\App\RouteHandlers;
+namespace WMDE\Fundraising\Frontend\App\Controllers;
 
 use Psr\Log\LogLevel;
 use Symfony\Component\HttpFoundation\ParameterBag;
@@ -14,41 +14,31 @@ use WMDE\Fundraising\Frontend\Factories\FunFunFactory;
 use WMDE\Fundraising\Frontend\Infrastructure\Payment\PayPalPaymentNotificationVerifierException;
 use WMDE\Fundraising\PaymentContext\RequestModel\PayPalPaymentNotificationRequest;
 
-/**
- * @license GNU GPL v2+
- * @author Kai Nissen < kai.nissen@wikimedia.de >
- */
-class PayPalNotificationHandler {
+class PaypalNotificationController {
 
 	private const MSG_NOT_HANDLED = 'PayPal request not handled';
 
-	private $ffFactory;
-
-	public function __construct( FunFunFactory $ffFactory ) {
-		$this->ffFactory = $ffFactory;
-	}
-
-	public function handle( Request $request ): Response {
+	public function handle( FunFunFactory $ffFactory, Request $request ): Response {
 		$post = $request->request;
 
 		if ( !$this->requestIsForPaymentCompletion( $post ) ) {
-			$this->ffFactory->getPaypalLogger()->log( LogLevel::INFO, self::MSG_NOT_HANDLED, [ 'post_vars' => $post->all() ] );
+			$ffFactory->getPaypalLogger()->log( LogLevel::INFO, self::MSG_NOT_HANDLED, [ 'post_vars' => $post->all() ] );
 			return new Response( '', Response::HTTP_OK );
 		}
 
 		try {
-			$this->ffFactory->getPayPalPaymentNotificationVerifier()->verify( $post->all() );
+			$ffFactory->getPayPalPaymentNotificationVerifier()->verify( $post->all() );
 		} catch ( PayPalPaymentNotificationVerifierException $e ) {
-			$this->ffFactory->getPaypalLogger()->log( LogLevel::ERROR, $e->getMessage(), [
-				'post_vars' => $request->request->all()
+			$ffFactory->getPaypalLogger()->log( LogLevel::ERROR, $e->getMessage(), [
+				'post_vars' => $post->all()
 			] );
 			return $this->createErrorResponse( $e );
 		}
 
-		$useCase = $this->ffFactory->newHandlePayPalPaymentCompletionNotificationUseCase( $this->getUpdateToken( $post ) );
+		$useCase = $ffFactory->newHandlePayPalPaymentCompletionNotificationUseCase( $this->getUpdateToken( $post ) );
 
 		$response = $useCase->handleNotification( $this->newUseCaseRequestFromPost( $post ) );
-		$this->logResponseIfNeeded( $response, $request );
+		$this->logResponseIfNeeded( $ffFactory, $response, $post );
 
 		return new Response( '', Response::HTTP_OK ); # PayPal expects an empty response
 	}
@@ -124,7 +114,7 @@ class PayPalNotificationHandler {
 		}
 	}
 
-	private function logResponseIfNeeded( PaypalNotificationResponse $response, Request $request ) {
+	private function logResponseIfNeeded( FunFunFactory $ffFactory, PaypalNotificationResponse $response, ParameterBag $post ) {
 		if ( $response->notificationWasHandled() ) {
 			return;
 		}
@@ -133,8 +123,8 @@ class PayPalNotificationHandler {
 		$message = $context['message'] ?? self::MSG_NOT_HANDLED;
 		$logLevel = $response->hasErrors() ? LogLevel::ERROR : LogLevel::INFO;
 		unset( $context['message'] );
-		$context['post_vars'] = $request->request->all();
-		$this->ffFactory->getPaypalLogger()->log( $logLevel, $message, $context );
+		$context['post_vars'] = $post->all();
+		$ffFactory->getPaypalLogger()->log( $logLevel, $message, $context );
 	}
 
 }
